Enable bullshit warnings so Xcode 5 quits warning me to "upgrade" my project settings. Fix said warnings by adding ugly casts.

+++ TLMSplitView.m
@@ -45,7 +45,7 @@
     BOOL inDivider = NO;
     NSPoint mouseLoc = [self convertPoint:[theEvent locationInWindow] fromView:nil];
     NSArray *subviews = [self subviews];
-    int i, count = [subviews count];
+    NSUInteger i, count = [subviews count];
     id view;
     NSRect divRect;
